Moved the abstracted api call files into the non-deprecated folder
