+++ src/com/company/api_calls_deprecated/individual/Liqui/AbstractLiqui.java
@@ -1,6 +1,6 @@
 package com.company.api_calls_deprecated.individual.Liqui;
 
-import com.company.api_calls_deprecated.AbstractJSONCaller;
+import com.company.api_calls.AbstractJSONCaller;
 import com.company.controller.ControllerInterface;
 
 /**
