+++ src/com/company/api_calls_deprecated/individual/ShapeShift/AbstractShapeShift.java
@@ -1,6 +1,6 @@
 package com.company.api_calls_deprecated.individual.ShapeShift;
 
-import com.company.api_calls_deprecated.AbstractJSONCaller;
+import com.company.api_calls.AbstractJSONCaller;
 import com.company.controller.ControllerInterface;
 
 /**
