+++ src/com/company/controller/controllers/main_controller/MainController.java
@@ -1,6 +1,6 @@
 package com.company.controller.controllers.main_controller;
 
-import com.company.api_calls_deprecated.APICallerInterface;
+import com.company.api_calls.APICallerInterface;
 import com.company.api_calls_deprecated.individual.CoinBase.*;
 import com.company.api_calls_deprecated.individual.CoinMarketCap.*;
 import com.company.tools_deprecated.Errors;
