+++ src/com/company/controller/controllers/main_controller/MainControllerInterface.java
@@ -1,6 +1,6 @@
 package com.company.controller.controllers.main_controller;
 
-import com.company.api_calls_deprecated.APICallerInterface;
+import com.company.api_calls.APICallerInterface;
 import com.company.tools_deprecated.FiatCurrencies;
 import com.company.controller.ControllerInterface;
